12555 - Support for Windows Platform
Add initial support for Microsoft Windows.

The main issue is that we hard coded \tmp in a few places
and used unix signal handlers that does not exist in windows

Windows has no Syslog and the syslog module for ruby is not
in the windows installers so we do a few tweaks to the syslog
logger and tests to just nil out in windows.  We should look
into adding a logger for the windows eventlog using win32-eventlog

In a few tests we used pwd and other unix utilities, we replace
all those to calls to a few ruby invocations that is more portable

All tests bar 1 pass, the remaining issue is in the PSK plugin where
it uses the Unix gid.