12555 - Support for Windows Platform
Windows dosbox doesn't show colors correctly so default to off on
that platform.

Figuring out the correct terminal dimension across the various ways
that windows can do terminals just isnt worth it so default to 80x40
for the purposes of creating the rpc progress bar

^C doesn't interrupt windows IO so we start a background thread that
just sleeps for a second, this gives the ruby VM a chance every second
to process ^C which would then exit.  We do this in both the application
wrapper and the main wrapper

The help template now defaults to being in the directory that the config
file lives in